classify: refactor to new file
